common interface to hyperbolic spaces
